--- .storybook/preview.ts
@@ -1,8 +1,0 @@
-export const parameters = {
-  controls: {
-    matchers: {
-      color: /(background|color)$/i,
-      date: /Date$/,
-    },
-  },
-}
