+++ .storybook/stories/Text.stories.tsx
@@ -1,6 +1,5 @@
 import { Meta, StoryObj } from '@storybook/react-native'
-import { withBackgrounds } from '@storybook/addon-ondevice-backgrounds'
-import { palette } from '../../src/features/ui/theme/palette'
+
 import React from 'react'
 import { View } from 'react-native'
 
@@ -9,27 +8,6 @@ import { Text, type IText } from '../../src/features/ui/components/Text'
 const meta: Meta<IText> = {
   title: 'AppText',
   component: Text,
-  parameters: {
-    backgrounds: {
-      default: 'warm',
-      values: [
-        { name: 'background', value: palette.springWood, default: true },
-        { name: 'white', value: 'white' },
-        { name: 'black', value: 'black' },
-      ],
-    },
-  },
-  decorators: [
-    withBackgrounds,
-    (Story) => (
-      <View
-        style={{
-          flex: 1,
-        }}>
-        <Story />
-      </View>
-    ),
-  ],
 }
 
 export default meta
@@ -45,9 +23,7 @@ export const Basic: Story = {
         padding: 20,
         paddingTop: 40,
       }}>
-      <Text variant="h1" {...restProps}>
-        Heading 1
-      </Text>
+      <Text variant="h1">Heading 1</Text>
       <Text variant="h2">Heading 2</Text>
       <Text variant="h2">Heading 3</Text>
       <Text variant="h3">Heading 4</Text>
